ai-control-plane/requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# AI Integration
anthropic==0.43.0

# Database
sqlalchemy[asyncio]==2.0.36
alembic==1.14.1
aiosqlite==0.20.0

# Validation
openapi-spec-validator==0.7.1
jsonschema==4.23.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Observability
structlog==24.4.0
prometheus-client==0.21.1

# Async / Files
aiofiles==24.1.0

# Configuration
pyyaml==6.0.2

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1

# Utilities
python-dotenv==1.0.1
